Split out ArtifactUtil & ArtifactHandler.
