+++ source/compiler-core/slang-artifact-desc-util.cpp
@@ -620,31 +620,6 @@ UnownedStringSlice ArtifactDescUtil::getDefaultExtension(const ArtifactDesc& des
     return getBaseNameFromPath(desc, path);
 }
 
-/* static */String ArtifactDescUtil::getBaseName(IArtifact* artifact)
-{
-    if (auto fileRep = findRepresentation<IFileArtifactRepresentation>(artifact))
-    {
-        return getBaseName(artifact->getDesc(), fileRep);
-    }
-    // Else use the name
-    return artifact->getName();
-}
-
-/* static */String ArtifactDescUtil::getParentPath(IFileArtifactRepresentation* fileRep)
-{
-    UnownedStringSlice path(fileRep->getPath());
-    return Path::getParentDirectory(path);
-}
-
-/* static */String ArtifactDescUtil::getParentPath(IArtifact* artifact)
-{
-    if (auto fileRep = findRepresentation<IFileArtifactRepresentation>(artifact))
-    {
-        return getParentPath(fileRep);
-    }
-    return String();
-}
-
 /* static */SlangResult ArtifactDescUtil::calcPathForDesc(const ArtifactDesc& desc, const UnownedStringSlice& basePath, StringBuilder& outPath)
 {
     outPath.Clear();
@@ -688,62 +663,6 @@ UnownedStringSlice ArtifactDescUtil::getDefaultExtension(const ArtifactDesc& des
     }
 
     return SLANG_OK;
-}
-
-/* static */ComPtr<IArtifactContainer> ArtifactDescUtil::createContainer(const ArtifactDesc& desc)
-{
-    const auto containerDesc = ArtifactDesc::make(ArtifactKind::Container, ArtifactPayload::CompileResults, desc.style);
-    return ArtifactContainer::create(containerDesc);
-}
-
-/* static */ComPtr<IArtifactContainer> ArtifactDescUtil::createResultsContainer()
-{
-    return ArtifactContainer::create(ArtifactDesc::make(ArtifactKind::Container, ArtifactPayload::CompileResults));
-}
-
-/* static */ComPtr<IArtifact> ArtifactDescUtil::createArtifactForCompileTarget(SlangCompileTarget target)
-{
-    auto desc = makeDescFromCompileTarget(target);
-    return Artifact::create(desc);
-}
-
-/* static */bool ArtifactDescUtil::isSignificant(IArtifact* artifact, void* data)
-{
-    SLANG_UNUSED(data);
-
-    const auto desc = artifact->getDesc();
-
-    // Containers are not significant as of themselves, they may contain something tho
-    if (isDerivedFrom(desc.kind, ArtifactKind::Container))
-    {
-        return false;
-    }
-
-    // If it has no payload.. we are done
-    if (desc.payload == ArtifactPayload::None ||
-        desc.payload == ArtifactPayload::Invalid)
-    {
-        return false;
-    }
-
-    // If it's binary like or assembly/source we it's significant
-    if (isDerivedFrom(desc.kind, ArtifactKind::BinaryLike) ||
-        desc.kind == ArtifactKind::Assembly ||
-        desc.kind == ArtifactKind::Source)
-    {
-        return true;
-    }
-
-    
-    /* Hmm, we might want to have a base class for 'signifiant' payloads, 
-    where signifiance here means somewhat approximately 'the meat' of a compilation result, 
-    as contrasted with 'meta data', 'diagnostics etc'*/
-    if (isDerivedFrom(desc.payload, ArtifactPayload::MetaData))
-    {
-        return false;
-    }
-
-    return true;
 }
 
 } // namespace Slang
