+++ source/compiler-core/slang-artifact-desc-util.h
@@ -75,12 +75,8 @@ struct ArtifactDescUtil
         /// If there is a path set, will extract the name from that (stripping prefix, extension as necessary).
         /// Else if there is an explicit name set, this is returned.
         /// Else returns the empty string
-    static String getBaseName(IArtifact* artifact);
+    
     static String getBaseName(const ArtifactDesc& desc, IFileArtifactRepresentation* fileRep);
-
-        /// Get the parent path (empty if there isn't one)
-    static String getParentPath(IArtifact* artifact);
-    static String getParentPath(IFileArtifactRepresentation* fileRep);
 
         /// Given a desc, and a basePath returns a suitable path for a entity of specified desc
     static SlangResult calcPathForDesc(const ArtifactDesc& desc, const UnownedStringSlice& basePath, StringBuilder& outPath);
@@ -90,20 +86,6 @@ struct ArtifactDescUtil
 
         /// Make ArtifactDesc from target
     static bool isDescDerivedFrom(const ArtifactDesc& desc, const ArtifactDesc& from);
-
-        /// Create an empty container which is compatible with the desc
-    static ComPtr<IArtifactContainer> createContainer(const ArtifactDesc& desc);
-
-        /// Create a generic container
-    static ComPtr<IArtifactContainer> createResultsContainer();
-
-        /// Creates an empty artifact for a type
-    static ComPtr<IArtifact> createArtifactForCompileTarget(SlangCompileTarget target);
-
-        /// Returns true if an artifact is 'significant'
-    static bool isSignificant(IArtifact* artifact, void* data = nullptr);
-        /// Find a significant artifact
-    static IArtifact* findSignificant(IArtifact* artifact) { return artifact->findRecursivelyByPredicate(&isSignificant, nullptr); }
 };
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
